+++ src/session.hpp
@@ -67,7 +67,6 @@ namespace sdk {
         void set_transaction_memo(const std::string& txhash_hex, const std::string& memo);
 
         nlohmann::json get_fee_estimates();
-        std::string get_mnemonic_passphrase(const std::string& password);
 
         std::string get_system_message();